Execution now executes with execve, command_not_found error now works when PATH is unset
