+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: eberkowi <eberkowi@student.hive.fi>        +#+  +:+       +#+         #
+#    By: maheleni <maheleni@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/09/20 15:07:12 by eberkowi          #+#    #+#              #
-#    Updated: 2024/10/11 10:23:41 by maheleni         ###   ########.fr        #
+#    Updated: 2024/10/11 11:56:30 by maheleni         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -29,6 +29,7 @@ CFILES := main.c \
 		parsing/expand_variables.c \
 		parsing/linked_list_utils.c \
 		built_ins/exit_command.c \
+		environment/list_add_and_remove.c \
 		environment/environment_utils.c \
 		environment/print_environment.c \
 		input_and_signals/handle_inputs.c \
